Add files via upload
You're building a situation for a pet care center that employs various people to care for
different types of pets.

In this pet care center, dogs and cats are the most common pets, and they are taken care of
by two types of employees: Veterinarians and Caretakers.

To simulate this situation, you need to implement an Object-Oriented program with the
following s situation:

1. Implement an abstract base class `Animal` with the following features: - A protected
string variable `name`. - A public constructor that initializes `name`. - A pure virtual
function `make Sound ()`.

2. Implement two derived classes `Dog` and `Cat` that inherit from `Animal`. Each of these
classes should override the `make Sound ()` function.

3. Implement two base classes `Person` and `Worker`. - `Person` should have protected
variables `name` (string) and `age` (int), and a public constructor that initializes these
variables. - `Worker` should have a protected variable `jobTitle` (string), and a public
constructor that initializes `jobTitle`.

4. Implement two classes `Veterinarian` and `Caretaker` that inherit from both `Person` and
`Worker`. - The `Veterinarian` and `Caretaker` classes should include a public constructor
that initializes all the variables from its base classes. - Add a function `work (Animal*
animal)` in both classes. This function should output a string indicating the worker's
jobTitle, their name, and the sound of the animal they are taking care of.

5. In your main function, create an array of pointers to `Animal` that includes instances of
`Dog` and `Cat`. Also, create a vector of `Worker` pointers that includes instances of
`Veterinarian` and `Caretaker`.